fixed ResNet50 model loading, landmark overlay alignment, and update comparison with accurate metrics

+++ web/src/lib/micro-trainer.ts
@@ -53,7 +53,7 @@ export function generateSyntheticData(
 
   return {
     xs: tf.tensor2d(shuffledData),
-    ys: tf.tensor1d(shuffledLabels, 'int32'),
+    ys: tf.tensor1d(shuffledLabels, 'float32'),
   };
 }
